requirements.txt
Django==1.10.5
django-bootstrap3==8.2.1
django-registration-redux==1.4
MySQL-python==1.2.5
